+++ resources/views/register.blade.php
@@ -4,6 +4,13 @@
   <meta charset="utf-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1">
+  <meta name="author" content="Alexis Saettler" />
+  <!--
+    This file is part of asbiin/laravel-webauthn project.
+
+    @copyright Alexis SAETTLER © 2019
+    @license MIT
+  -->
 
   <title>{{ config('app.name', 'Laravel') }}</title>